refactor & clean up
